Build 2.12 		1.  Extended support for OpenSUSE,FreeBSD and Linux Mint. 		2.  Renamed the "Job_Termination_Script.pl" to "job_termination.pl" with new design. 		3.  Fixed restore issue when directory name in restore set file having white-space. 		4.  Fixed the scheduler issue when crontab path is link. 		5.  Fixed the status retrival issue when the username having character "cd". 		6.  Fixed the transfer rate issue in Backup/Restore progress bar. 		7.  Fixed the linux binary finding issue in "Account_Setting.pl" script when zip file passed as argument. 		8.  Fixed the duplicate header display issue in Restore_Version.pl script. 		9.  Fixed the IDrive account validation issue with fallback logic.

+++ Login.pl
@@ -15,13 +15,7 @@ my $isPrivate = 0;
 my $encType = undef;
 my $pvtKey = undef;
 my $pvtKeyField = undef;
-#Check if EVS Binary exists.
-my $err_string = checkBinaryExists();
-if($err_string ne "") {
-        print qq($err_string);
-        exit 1;
-}
-
+my $accDetailsCheck = 0;
 headerDisplay($0);
 my $CurrentUser = getCurrentUser();
 if ($CurrentUser ne ""){
@@ -30,7 +24,8 @@ if ($CurrentUser ne ""){
 }
 print $lineFeed.Constants->CONST->{'displayUserMessage'}->('Enter your',$appType,'username: '); 
 $userName = <STDIN>;
-$userName =~ s/^[\s\t]+|[\s\t]+$//g;
+$userName =~ s/^[\s\t]+//;
+$userName =~ s/[\s\t]+$//;
 checkInput(\$userName);
 unless (validateUserName($userName)){
       print Constants->CONST->{'InvalidUserPattern'}.$lineFeed;
@@ -39,7 +34,10 @@ unless (validateUserName($userName)){
 my $userDir = "$usrProfilePath/$userName";
 $confFilePath = "$userDir/".Constants->CONST->{'configurationFile'};
 checkIfDirExits($confFilePath);
-
+if (! checkIfEvsWorking($dedup)){
+        print Constants->CONST->{'EvsProblem'}.$lineFeed;
+        exit 0;
+}
 if(getAccountConfStatus($confFilePath)){
 	exit(0);
 }
@@ -80,15 +78,24 @@ $password = <STDIN>;
 system('stty','echo');
 chomp($password);
 $password =~ s/^\s+|\s+$//g;
-
-if(length($password) > 20) {
-	print $lineFeed.Constants->CONST->{'InvalidUnamePwd'}.$lineFeed;
-	exit;
+unless (validatePassword($password)){
+	print $lineFeed.Constants->CONST->{'InvalidPassPattern'}.$lineFeed;
+      	exit 0;
 }
-createEncodeFile($password, $pwdPath);
-createEncodeSecondaryFile($password, $enPwdPath, $userName);
-validateAccount();
-getQuota();
+CHECK_ACC_DETAILS:
+getAccountInfo();
+my $serverAddress = verifyAndLoadServerAddr();
+if ($serverAddress == 0){
+	cancelProcess();
+}
+createPasswordFiles($password,$pwdPath,$userName,$enPwdPath);
+setAccount($cnfgstat,\$pvt,$pvtPath);
+if ($message eq 'SUCCESS'){
+	getPvtKey();
+    verifyAccount();
+	checkAndUpdateClientRecord($userName,$password);
+}
+getQuotaForAccountSettings($accountQuota, $quotaUsed);
 #****************************************************************************************************
 # Subroutine Name         : validateAccount.
 # Objective               : This subroutine validates an user account if the account is
@@ -97,16 +104,15 @@ getQuota();
 #*****************************************************************************************************/
 sub validateAccount
 {	
-	my $reloginStatus = shift;
-	print $lineFeed.$lineFeed.Constants->CONST->{'verifyAccount'}.$lineFeed;
-	my $validateUtf8File = getOperationFile($validateOp);
+#	print $lineFeed.$lineFeed.Constants->CONST->{'verifyAccount'}.$lineFeed;
+	my $validateUtf8File = getOperationFile(Constants->CONST->{'ValidateOp'});
 	chomp($validateUtf8File);
 
 	#log API in trace file as well
 	traceLog("$lineFeed validateAccount: ", __FILE__, __LINE__);
 	$validateUtf8File =~ s/\'/\'\\''/g;
 	
-	my $idevsutilCommandLine = $idevsutilBinaryPath.$whiteSpace.$idevsutilArgument.$assignmentOperator."'".$validateUtf8File."'".$whiteSpace.$errorRedirection;
+	my $idevsutilCommandLine = "'$idevsutilBinaryPath'".$whiteSpace.$idevsutilArgument.$assignmentOperator."'".$validateUtf8File."'".$whiteSpace.$errorRedirection;
 	my $commandOutput = `$idevsutilCommandLine`;
 	traceLog("$lineFeed $commandOutput $lineFeed", __FILE__, __LINE__);
 	#print $tHandle "$lineFeed $commandOutput $lineFeed";
@@ -144,16 +150,29 @@ sub validateAccount
 		exit;
 	} elsif($commandOutput =~ m/bad response from proxy/) {
 		if ($reloginStatus ne 'on'){
-			print Constants->CONST->{'InvProxy'};
-			getProxyDetails();
-			putParameterValue(\"PROXY",\"$proxyStr",$confFilePath);
-			getConfigHashValue();
-			if($proxyStr eq ""){
-				$proxyStr = getProxy();
+			if ($accDetailsCheck == 1){
+				print Constants->CONST->{ProxyErr}.$lineFeed;
+                                exit(0);
 			}
-			if ($proxyStr ne ""){
+			print Constants->CONST->{'issueWithProxy'};
+			putParameterValue(\"PROXY",\"",$confFilePath);
+			if (getProxyDetails()){
+				putParameterValue(\"PROXY",\"$proxyStr",$confFilePath);
+				getConfigHashValue();
+				if($proxyStr eq ""){
+					$proxyStr = getProxy();
+				}
+				if ($proxyStr ne ""){
+					$commandOutput = '';
+					createPasswordFiles($password,$pwdPath,$userName,$enPwdPath);
+					$res = validateAccount('on');
+					return $res;
+				}
+			}else{
 				$commandOutput = '';
-				validateAccount('on');
+                                createPasswordFiles($password,$pwdPath,$userName,$enPwdPath);
+				$accDetailsCheck = 1;
+				goto CHECK_ACC_DETAILS;
 			}
 		}else{	
 			print Constants->CONST->{'InvProxy'}.$lineFeed;
@@ -167,6 +186,10 @@ sub validateAccount
 		}
 	}elsif($commandOutput =~ /Unable to reach the server\; account validation has failed/i){
 		if ($reloginStatus ne 'on'){
+			 if ($accDetailsCheck == 1){
+				print Constants->CONST->{ProxyErr}.$lineFeed;
+				exit(0);
+                         }
 			 print Constants->CONST->{'InvProxy'};
 			 getProxyDetails();
 			 putParameterValue(\"PROXY",\"$proxyStr",$confFilePath);
@@ -176,17 +199,19 @@ sub validateAccount
  			 }
 			 if ($proxyStr ne ""){
 			        $commandOutput = '';
-				validateAccount('on');
+				createPasswordFiles($password,$pwdPath,$userName,$enPwdPath);
+				$res = validateAccount('on');
+				return $res;
 			 }
 		}else{
 			print qq($lineFeed$& $lineFeed);
 			print Constants->CONST->{ProxyErr}.$lineFeed;
 			if(-e $enPwdPath){
-        	                unlink($enPwdPath);
-	                }
-                	if(-e $pwdPath){
-                        	unlink($pwdPath);
-	                }
+				unlink($enPwdPath);
+			}
+			if(-e $pwdPath){
+				unlink($pwdPath);
+			}
 			exit(0);
 		}
 	}else {
@@ -199,9 +224,7 @@ sub validateAccount
 		print Constants->CONST->{ProxyErr}.$lineFeed;
 		exit(0);	
 	}
-	if ($reloginStatus ne 'on'){
-		verifyAccount();
-	}
+	return $commandOutput;
 }
 
 #****************************************************************************************************
@@ -219,11 +242,11 @@ sub configureAccount() {
 	traceLog("$lineFeed configureAccount: ", __FILE__, __LINE__);
 	#print $tHandle "$lineFeed configureAccount: ";
 	
-	my $configUtf8File = getOperationFile($configOp, $encType);
+	my $configUtf8File = getOperationFile(Constants->CONST->{'ConfigOp'}, $encType);
 	chomp($configUtf8File);
 	$configUtf8File =~ s/\'/\'\\''/g;
 	
-	$idevsutilCommandLine = $idevsutilBinaryPath.$whiteSpace.$idevsutilArgument.$assignmentOperator."'".$configUtf8File."'".$whiteSpace.$errorRedirection;
+	$idevsutilCommandLine = "'$idevsutilBinaryPath'".$whiteSpace.$idevsutilArgument.$assignmentOperator."'".$configUtf8File."'".$whiteSpace.$errorRedirection;
 	my $commandOutput = `$idevsutilCommandLine`;
 	traceLog("$lineFeed $commandOutput $lineFeed", __FILE__, __LINE__);
 	#print $tHandle "$lineFeed $commandOutput $lineFeed";
@@ -236,33 +259,39 @@ sub configureAccount() {
 # Objective               : This subroutine varifies the encryption key by creating Folder.
 # Added By                : Dhritikana
 #*****************************************************************************************************/
-sub verifyAccount()
+sub verifyAccount
 {
 	if($isPrivate) {
 		createEncodeFile($pvtKey, $pvtPath);
 	}
 	#get evs server address for other APIs
 	getServerAddr();
+	my $retType = '';
 	if($isPrivate eq 1) {
 		print $lineFeed.Constants->CONST->{'verifyPvt'}.$lineFeed;
-		my $pvtVerifyUtfFile = getOperationFile($verifyPvtOp);
-		my $tmp_pvtVerifyUtfFile = $pvtVerifyUtfFile;
-		$tmp_pvtVerifyUtfFile =~ s/\'/\'\\''/g;
-		my $tmp_idevsutilBinaryPath = $idevsutilBinaryPath;
-		$tmp_idevsutilBinaryPath =~ s/\'/\'\\''/g;
+		if ($dedup eq 'off'){
+			my $pvtVerifyUtfFile = getOperationFile(Constants->CONST->{'validatePvtKeyOp'});
+			my $tmp_pvtVerifyUtfFile = $pvtVerifyUtfFile;
+			$tmp_pvtVerifyUtfFile =~ s/\'/\'\\''/g;
+			my $tmp_idevsutilBinaryPath = $idevsutilBinaryPath;
+			$tmp_idevsutilBinaryPath =~ s/\'/\'\\''/g;
 		
-		my $idevsUtilCommand = "\'$tmp_idevsutilBinaryPath\'".$whiteSpace.$idevsutilArgument.$assignmentOperator."\'$tmp_pvtVerifyUtfFile\'".$whiteSpace.$errorRedirection;
-		my $retType = `$idevsUtilCommand`;
-		chomp($retType);
-		unlink($pvtVerifyUtfFile);
-		if($retType !~ /verification success/) {
+			my $idevsUtilCommand = "\'$tmp_idevsutilBinaryPath\'".$whiteSpace.$idevsutilArgument.$assignmentOperator."\'$tmp_pvtVerifyUtfFile\'".$whiteSpace.$errorRedirection;
+			$retType = `$idevsUtilCommand`;
+			chomp($retType);
+			unlink($pvtVerifyUtfFile);
+		}else{
+			$retType = getDeviceList();
+		}
+		if($retType =~ /encryption verification failed|Unable to proceed; private encryption key must be between 4 and 256 characters in length/) {
 #			print Constants->CONST->{'AskCorrectPvt'}.$lineFeed;
 			print Constants->CONST->{InvalidPvtKey}.' '.Constants->CONST->{TryAgain}.$lineFeed;
 			unlink($pwdPath);
 			unlink($pvtPath);
 			unlink($pvtPath."_SCH");
 		} 
-		elsif($retType =~ /verification success/) {
+	#	elsif($retType =~ /verification success|connection established/) {
+		else{
 			#Create Cache Directory 
 			createCache();
 			updateConf();
@@ -284,10 +313,10 @@ sub verifyAccount()
 #*****************************************************************************************************/
 sub updateConf()
 {
-	$dummyString = "XXXXX";
+	#$dummyString = "XXXXX";
 	$schPwdPath = "$pwdPath"."_SCH";
 	copy($pwdPath, $schPwdPath);
-#	putParameterValue(\"PASSWORD", \$dummyString, $confFilePath);
+	putParameterValue(\"DEDUP", \$dedup, $confFilePath);
 	putParameterValue(\"USERNAME", \$userName, $confFilePath);
 	if($isPrivate) {
 		$schPvtPath = $pvtPath."_SCH";
@@ -311,4 +340,125 @@ sub checkIfDirExits {
 		print Constants->CONST->{'loginConfigAgain'}.$lineFeed;
 		exit;
 	}
+}
+#****************************************************************************************************
+# Subroutine Name         : getAccountInfo.
+# Objective               : Gets the user account information by using CGI.
+# Added By                : Dhritikana.
+#*****************************************************************************************************/
+sub getAccountInfo {
+	print $lineFeed.Constants->CONST->{'verifyAccount'}.$lineFeed;
+	my $PATH = undef;
+	if($appType eq "IDrive") {
+		$PATH = $IDriveAccVrfLink;
+	} elsif($appType eq "IBackup") {
+		$PATH = $IBackupAccVrfLink;
+	}
+	
+	my $encodedUname = $userName;
+	my $encodedPwod = $password;
+	#URL DATA ENCODING#
+	foreach ($encodedUname, $encodedPwod) {
+		$_ =~ s/([^A-Za-z0-9])/sprintf("%%%02X", ord($1))/seg;
+	}
+	my $data = 'username='.$encodedUname.'&password='.$encodedPwod;
+	my $curl = `which curl`;
+	Chomp(\$curl);
+	if($proxyOn eq 1) {
+		$curlCmd = "$curl --max-time 15 -x http://$proxyIp:$proxyPort --proxy-user $proxyUsername:$proxyPassword -L -s -k -d '$data' '$PATH'";
+	} else {
+		$curlCmd = "$curl --max-time 15 -s -k -d '$data' '$PATH'";
+	}
+	my $res = `$curlCmd`;
+	if($res =~ /FAILURE/) {
+		if($res =~ /passwords do not match|Username or Password not found|invalid value passed for username|password too short|username too short|password too long|username too long/i) {
+#			undef $userName; Any specific reason behind writing this statement.
+			print ucfirst($&).". ".Constants->CONST->{'TryAgain'}."$lineFeed";
+			traceLog("$linefeed $curl failed, Reason: $&", __FILE__, __LINE__);
+#			removeFilesFolders(["$usrProfilePath/$userName"]);
+			cancelProcess();
+		}
+	} elsif ($res =~ /SSH-2.0-OpenSSH_6.8p1-hpn14v6|Protocol mismatch/) {
+		undef $userName;
+		print "$res.\n";
+		traceLog("$linefeed $curl failed, Reason: $res\n", __FILE__, __LINE__);
+		cancelProcess();
+	}
+	traceLog("curl res :$res", __FILE__, __LINE__);
+
+        if(!$res) {
+        	$res =	validateAccount();
+        } elsif( $res =~ /Unauthorized/) {
+            $res = validateAccount();
+        }
+        if ($res eq ''){
+           $res =  validateAccount();
+        }
+	
+	my %evsLoginHashOutput = parseXMLOutput(\$res);
+	chomp(%evsLoginHashOutput);
+	$encType = $evsLoginHashOutput{"enctype"} ne "" ? $evsLoginHashOutput{"enctype"} : $evsLoginHashOutput{"configtype"};
+	$plan_type = $evsLoginHashOutput{"plan_type"};
+	$message = $evsLoginHashOutput{"message"};
+	$cnfgstat = $evsLoginHashOutput{"cnfgstat"} ne "" ? $evsLoginHashOutput{"cnfgstat"} : $evsLoginHashOutput{"configstatus"};
+	$desc = $evsLoginHashOutput{"desc"};
+	$dedup = $evsLoginHashOutput{"dedup"} if($appType eq "IDrive");
+	$accountQuota = $evsLoginHashOutput{"quota"};
+	$quotaUsed = $evsLoginHashOutput{"quota_used"};
+	getServerAddr($evsLoginHashOutput{evssrvrip});	
+	#dedup check
+	if ($dedup eq 'on'){
+		$idevsutilBinaryName = "idevsutil_dedup";#Name of idevsutil binary#
+		$idevsutilBinaryPath = "$idriveServicePath/idevsutil_dedup";#Path of idevsutil binary#
+	}		
+}
+#****************************************************************************************************
+# Subroutine Name         : getPvtKey.
+# Objective               : Get the input for private key from user.
+# Added By                : Dhritikana
+#*****************************************************************************************************/
+sub getPvtKey{
+	if($cnfgstat eq "SET") {
+	# For Private Account Verifies the Private Encryption Key
+		if($encType eq "PRIVATE") {
+			print $lineFeed.Constants->CONST->{'AskPvtSetAcc'};
+			system('stty','-echo');
+			$pvtKey = getInput();
+			checkInput(\$pvtKey,$lineFeed);
+			system('stty','echo');
+			$isPrivate = 1;
+		}
+	}
+}
+
+#***************************************************************************************
+# Subroutine Name         : getMenuChoice
+# Objective               : get Menu choioce to check if user wants to configure his/her
+#                                                       with Default or Private Key.
+# Added By                : Dhritikana
+#****************************************************************************************/
+sub getMenuChoice {
+        my $count = 0;
+        while(!defined $menuChoice) {
+                if ($count < 4){
+                        $count++;
+                        print Constants->CONST->{'EnterChoice'};
+                        $menuChoice = <STDIN>;
+                        chomp $menuChoice;
+                        $menuChoice =~ s/^\s+|\s+$//;
+                        if($menuChoice =~ m/^\d$/) {
+                                if($menuChoice < 1 || $menuChoice > 2) {
+                                $menuChoice = undef;
+                                print Constants->CONST->{'InvalidChoice'}.$whiteSpace if ($count < 4);
+                        }
+                        }else {
+                                $menuChoice = undef;
+                                print Constants->CONST->{'InvalidChoice'}.$whiteSpace if ($count < 4);
+                        }
+                }else{
+                        print Constants->CONST->{'maxRetry'}.$lineFeed;
+                        $menuChoice='';
+                        exit;
+                }
+        }
 }
